+++ netbox_certificates/migrations/0018_service_tags_related_name.py
@@ -1,0 +1,28 @@
+import extras.managers
+import taggit.managers
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("netbox_certificates", "0017_service_default_related_name"),
+    ]
+
+    operations = [
+        migrations.SeparateDatabaseAndState(
+            state_operations=[
+                migrations.AlterField(
+                    model_name="service",
+                    name="tags",
+                    field=taggit.managers.TaggableManager(
+                        through="extras.TaggedItem",
+                        ordering=("weight", "name"),
+                        manager=extras.managers.NetBoxTaggableManager,
+                        related_name="netbox_certificates_service_tagged+",
+                    ),
+                ),
+            ],
+            database_operations=[],
+        ),
+    ]